[refactor]: split `iroha_torii` from `iroha`
