+++ Cargo.toml
@@ -16,7 +16,8 @@ categories = ["cryptography::cryptocurrencies"]
 [workspace.dependencies]
 iroha = { path = "cli" }
 iroha_dsl = { version = "=2.0.0-pre-rc.20", path = "dsl" }
-iroha_cli_derive = { version = "=2.0.0-pre-rc.20", path = "cli/derive" }
+iroha_torii = { version = "=2.0.0-pre-rc.20", path = "torii" }
+iroha_torii_derive = { version = "=2.0.0-pre-rc.20", path = "torii/derive" }
 iroha_macro_utils = { version = "=2.0.0-pre-rc.20", path = "macro/utils" }
 iroha_telemetry = { version = "=2.0.0-pre-rc.20", path = "telemetry" }
 iroha_telemetry_derive = { version = "=2.0.0-pre-rc.20", path = "telemetry/derive" }
@@ -205,7 +206,6 @@ clippy.wildcard_dependencies = "deny"
 resolver = "2"
 members = [
   "cli",
-  "cli/derive",
   "client",
   "client_cli",
   "config",
@@ -245,6 +245,8 @@ members = [
   "tools/swarm",
   "tools/wasm_builder_cli",
   "tools/wasm_test_runner",
+  "torii",
+  "torii/derive",
   "version",
   "version/derive",
   "wasm_codec",
