+++ torii/derive/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "iroha_cli_derive"
+name = "iroha_torii_derive"
 
 edition.workspace = true
 version.workspace = true
