+++ torii/src/routing.rs
@@ -5,9 +5,10 @@
 // FIXME: This can't be fixed, because one trait in `warp` is private.
 #![allow(opaque_hidden_inferred_bound)]
 
+#[cfg(feature = "telemetry")]
 use eyre::{eyre, WrapErr};
 use futures::TryStreamExt;
-use iroha_config::{client_api::ConfigurationDTO, torii::uri};
+use iroha_config::client_api::ConfigurationDTO;
 use iroha_core::{
     query::{pagination::Paginate, store::LiveQueryStoreHandle},
     smartcontracts::query::ValidQueryRequest,
@@ -33,7 +34,7 @@ use super::*;
 use crate::stream::{Sink, Stream};
 
 /// Filter for warp which extracts [`http::ClientQueryRequest`]
-fn client_query_request(
+pub fn client_query_request(
 ) -> impl warp::Filter<Extract = (http::ClientQueryRequest,), Error = warp::Rejection> + Copy {
     body::versioned::<SignedQuery>()
         .and(sorting())
@@ -64,7 +65,7 @@ fn cursor() -> impl warp::Filter<Extract = (ForwardCursor,), Error = warp::Rejec
 }
 
 /// Filter for warp which extracts pagination
-fn paginate() -> impl warp::Filter<Extract = (Pagination,), Error = warp::Rejection> + Copy {
+pub fn paginate() -> impl warp::Filter<Extract = (Pagination,), Error = warp::Rejection> + Copy {
     warp::query()
 }
 
@@ -74,7 +75,7 @@ fn fetch_size() -> impl warp::Filter<Extract = (FetchSize,), Error = warp::Rejec
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_transaction(
+pub async fn handle_transaction(
     queue: Arc<Queue>,
     sumeragi: SumeragiHandle,
     transaction: SignedTransaction,
@@ -98,13 +99,13 @@ async fn handle_transaction(
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_queries(
+pub async fn handle_queries(
     live_query_store: LiveQueryStoreHandle,
     sumeragi: SumeragiHandle,
 
     query_request: http::ClientQueryRequest,
 ) -> Result<Scale<BatchedResponse<Value>>> {
-    let handle = tokio::task::spawn_blocking(move || match query_request.0 {
+    let handle = task::spawn_blocking(move || match query_request.0 {
         QueryRequest::Query(QueryWithParameters {
             query: signed_query,
             sorting,
@@ -134,18 +135,18 @@ enum Health {
     Healthy,
 }
 
-fn handle_health() -> Json {
+pub fn handle_health() -> Json {
     reply::json(&Health::Healthy)
 }
 
 #[iroha_futures::telemetry_future]
-#[cfg(feature = "schema-endpoint")]
-async fn handle_schema() -> Json {
+#[cfg(feature = "schema")]
+pub async fn handle_schema() -> Json {
     reply::json(&iroha_schema_gen::build_schemas())
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_pending_transactions(
+pub async fn handle_pending_transactions(
     queue: Arc<Queue>,
     sumeragi: SumeragiHandle,
     pagination: Pagination,
@@ -164,13 +165,13 @@ async fn handle_pending_transactions(
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_get_configuration(kiso: KisoHandle) -> Result<Json> {
+pub async fn handle_get_configuration(kiso: KisoHandle) -> Result<Json> {
     let dto = kiso.get_dto().await?;
     Ok(reply::json(&dto))
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_post_configuration(
+pub async fn handle_post_configuration(
     kiso: KisoHandle,
     value: ConfigurationDTO,
 ) -> Result<impl Reply> {
@@ -179,7 +180,7 @@ async fn handle_post_configuration(
 }
 
 #[iroha_futures::telemetry_future]
-async fn handle_blocks_stream(kura: Arc<Kura>, mut stream: WebSocket) -> eyre::Result<()> {
+pub async fn handle_blocks_stream(kura: Arc<Kura>, mut stream: WebSocket) -> eyre::Result<()> {
     let BlockSubscriptionRequest(mut from_height) = stream.recv().await?;
 
     let mut interval = tokio::time::interval(std::time::Duration::from_millis(10));
@@ -221,7 +222,7 @@ async fn handle_blocks_stream(kura: Arc<Kura>, mut stream: WebSocket) -> eyre::R
     }
 }
 
-mod subscription {
+pub mod subscription {
     //! Contains the `handle_subscription` functions and used for general routing.
 
     use super::*;
@@ -303,7 +304,7 @@ mod subscription {
 
 #[iroha_futures::telemetry_future]
 #[cfg(feature = "telemetry")]
-async fn handle_version(sumeragi: SumeragiHandle) -> Json {
+pub async fn handle_version(sumeragi: SumeragiHandle) -> Json {
     use iroha_version::Version;
 
     let string = sumeragi
@@ -315,7 +316,7 @@ async fn handle_version(sumeragi: SumeragiHandle) -> Json {
 }
 
 #[cfg(feature = "telemetry")]
-fn handle_metrics(sumeragi: &SumeragiHandle) -> Result<String> {
+pub fn handle_metrics(sumeragi: &SumeragiHandle) -> Result<String> {
     if let Err(error) = sumeragi.update_metrics() {
         iroha_logger::error!(%error, "Error while calling sumeragi::update_metrics.");
     }
@@ -333,7 +334,7 @@ fn update_metrics_gracefully(sumeragi: &SumeragiHandle) {
 
 #[cfg(feature = "telemetry")]
 #[allow(clippy::unnecessary_wraps)]
-fn handle_status(
+pub fn handle_status(
     sumeragi: &SumeragiHandle,
     accept: Option<impl AsRef<str>>,
     tail: &warp::path::Tail,
@@ -368,208 +369,5 @@ fn handle_status(
             .map(|segment| reply::json(segment).into_response())?;
 
         Ok(reply)
-    }
-}
-
-impl Torii {
-    /// Construct `Torii`.
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        kiso: KisoHandle,
-        config: &ToriiConfiguration,
-        queue: Arc<Queue>,
-        events: EventsSender,
-        notify_shutdown: Arc<Notify>,
-        sumeragi: SumeragiHandle,
-        query_service: LiveQueryStoreHandle,
-        kura: Arc<Kura>,
-    ) -> Self {
-        Self {
-            kiso,
-            queue,
-            events,
-            notify_shutdown,
-            sumeragi,
-            query_service,
-            kura,
-            address: config.api_url.clone(),
-            transaction_max_content_length: config.max_content_len.into(),
-        }
-    }
-
-    /// Helper function to create router. This router can tested without starting up an HTTP server
-    #[allow(clippy::too_many_lines)]
-    fn create_api_router(&self) -> impl warp::Filter<Extract = impl warp::Reply> + Clone + Send {
-        let health_route = warp::get()
-            .and(warp::path(uri::HEALTH))
-            .and_then(|| async { Ok::<_, Infallible>(handle_health()) });
-
-        let get_router = warp::get().and(
-            endpoint3(
-                handle_pending_transactions,
-                warp::path(uri::PENDING_TRANSACTIONS)
-                    .and(add_state!(self.queue, self.sumeragi,))
-                    .and(paginate()),
-            )
-            .or(warp::path(uri::CONFIGURATION)
-                .and(add_state!(self.kiso))
-                .and_then(|kiso| async move {
-                    Ok::<_, Infallible>(WarpResult(handle_get_configuration(kiso).await))
-                })),
-        );
-
-        let get_router_status = warp::path(uri::STATUS)
-            .and(add_state!(self.sumeragi.clone()))
-            .and(warp::header::optional(warp::http::header::ACCEPT.as_str()))
-            .and(warp::path::tail())
-            .and_then(|sumeragi, accept: Option<String>, tail| async move {
-                Ok::<_, Infallible>(WarpResult(handle_status(&sumeragi, accept.as_ref(), &tail)))
-            });
-        let get_router_metrics = warp::path(uri::METRICS)
-            .and(add_state!(self.sumeragi))
-            .and_then(|sumeragi| async move {
-                Ok::<_, Infallible>(WarpResult(handle_metrics(&sumeragi)))
-            });
-        let get_api_version = warp::path(uri::API_VERSION)
-            .and(add_state!(self.sumeragi.clone()))
-            .and_then(|sumeragi| async { Ok::<_, Infallible>(handle_version(sumeragi).await) });
-
-        #[cfg(feature = "telemetry")]
-        let get_router = get_router.or(warp::any()
-            .and(get_router_status)
-            .or(get_router_metrics)
-            .or(get_api_version));
-
-        #[cfg(feature = "schema-endpoint")]
-        let get_router = get_router.or(warp::path(uri::SCHEMA)
-            .and_then(|| async { Ok::<_, Infallible>(handle_schema().await) }));
-
-        let post_router = warp::post()
-            .and(
-                endpoint3(
-                    handle_transaction,
-                    warp::path(uri::TRANSACTION)
-                        .and(add_state!(self.queue, self.sumeragi))
-                        .and(warp::body::content_length_limit(
-                            self.transaction_max_content_length,
-                        ))
-                        .and(body::versioned()),
-                )
-                .or(endpoint3(
-                    handle_queries,
-                    warp::path(uri::QUERY)
-                        .and(add_state!(self.query_service, self.sumeragi,))
-                        .and(client_query_request()),
-                ))
-                .or(endpoint2(
-                    handle_post_configuration,
-                    warp::path(uri::CONFIGURATION)
-                        .and(add_state!(self.kiso))
-                        .and(warp::body::json()),
-                )),
-            )
-            .recover(|rejection| async move { body::recover_versioned(rejection) });
-
-        let events_ws_router = warp::path(uri::SUBSCRIPTION)
-            .and(add_state!(self.events))
-            .and(warp::ws())
-            .map(|events, ws: Ws| {
-                ws.on_upgrade(|this_ws| async move {
-                    if let Err(error) = subscription::handle_subscription(events, this_ws).await {
-                        iroha_logger::error!(%error, "Failure during subscription");
-                    }
-                })
-            });
-
-        // `warp` panics if there is `/` in the string given to the `warp::path` filter
-        // Path filter has to be boxed to have a single uniform type during iteration
-        let block_ws_router_path = uri::BLOCKS_STREAM
-            .split('/')
-            .skip_while(|p| p.is_empty())
-            .fold(warp::any().boxed(), |path_filter, path| {
-                path_filter.and(warp::path(path)).boxed()
-            });
-
-        let blocks_ws_router = block_ws_router_path
-            .and(add_state!(self.kura))
-            .and(warp::ws())
-            .map(|sumeragi: Arc<_>, ws: Ws| {
-                ws.on_upgrade(|this_ws| async move {
-                    if let Err(error) = handle_blocks_stream(sumeragi, this_ws).await {
-                        iroha_logger::error!(%error, "Failed to subscribe to blocks stream");
-                    }
-                })
-            });
-
-        let ws_router = events_ws_router.or(blocks_ws_router);
-
-        warp::any()
-            .and(
-                // we want to avoid logging for the "health" endpoint.
-                // we have to place it **first** so that warp's trace will
-                // not log 404 if it doesn't find "/health" which might be placed
-                // **after** `.with(trace)`
-                health_route,
-            )
-            .or(ws_router
-                .or(get_router)
-                .or(post_router)
-                .with(warp::trace::request()))
-    }
-
-    /// Start main api endpoints.
-    ///
-    /// # Errors
-    /// Can fail due to listening to network or if http server fails
-    fn start_api(self: Arc<Self>) -> eyre::Result<Vec<tokio::task::JoinHandle<()>>> {
-        let torii_address = &self.address;
-
-        let mut handles = vec![];
-        match torii_address.to_socket_addrs() {
-            Ok(addrs) => {
-                for addr in addrs {
-                    let torii = Arc::clone(&self);
-
-                    let api_router = torii.create_api_router();
-                    let signal_fut = async move { torii.notify_shutdown.notified().await };
-                    let (_, serve_fut) =
-                        warp::serve(api_router).bind_with_graceful_shutdown(addr, signal_fut);
-
-                    handles.push(task::spawn(serve_fut));
-                }
-
-                Ok(handles)
-            }
-            Err(error) => {
-                iroha_logger::error!(%torii_address, %error, "API address configuration parse error");
-                Err(eyre::Error::new(error))
-            }
-        }
-    }
-
-    /// To handle incoming requests `Torii` should be started first.
-    ///
-    /// # Errors
-    /// Can fail due to listening to network or if http server fails
-    #[iroha_futures::telemetry_future]
-    pub(crate) async fn start(self) -> eyre::Result<()> {
-        let torii = Arc::new(self);
-        let mut handles = vec![];
-
-        handles.extend(Arc::clone(&torii).start_api()?);
-
-        handles
-            .into_iter()
-            .collect::<FuturesUnordered<_>>()
-            .for_each(|handle| {
-                if let Err(error) = handle {
-                    iroha_logger::error!(%error, "Join handle error");
-                }
-
-                futures::future::ready(())
-            })
-            .await;
-
-        Ok(())
     }
 }
